+++ src/main/java/com/ssg/usms/business/user/controller/UserController.java
@@ -1,21 +1,20 @@
 package com.ssg.usms.business.user.controller;
 
 
+import com.ssg.usms.business.security.login.UsmsUserDetails;
 import com.ssg.usms.business.user.dto.HttpRequestModifyUserDto;
 import com.ssg.usms.business.user.dto.HttpRequestSignUpDto;
 import com.ssg.usms.business.user.dto.HttpResponseUserDto;
 import com.ssg.usms.business.user.dto.SecurityState;
-import com.ssg.usms.business.user.exception.NotAllowedKeyExcetpion;
 import com.ssg.usms.business.user.exception.NotAllowedSecondPasswordException;
-import com.ssg.usms.business.user.service.SignUpService;
 import com.ssg.usms.business.user.service.UserService;
 import com.ssg.usms.business.user.util.JwtUtil;
-import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,33 +27,31 @@ import static com.ssg.usms.business.user.constant.UserConstants.*;
 @Validated
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api")
 public class UserController {
 
     private final JwtUtil jwtUtil;
 
     private final UserService userService;
-    private final SignUpService signUpService;
 
-    @GetMapping("/users")
+    @GetMapping("/api/users")
     public String users(){
         return "users ADMIN";
     }
 
-    @PostMapping("/users")
+    @PostMapping("/api/users")
     public ResponseEntity<Void> SignUp(@Valid @RequestBody HttpRequestSignUpDto httpRequestSignUpDto, HttpServletRequest request){
 
         String authorization = request.getHeader(HttpHeaders.AUTHORIZATION);
 
         jwtUtil.VerifyToken(authorization,"Identification");
 
-        signUpService.SignUp(httpRequestSignUpDto);
+        userService.SignUp(httpRequestSignUpDto);
 
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
 
-    @GetMapping("/user")
+    @GetMapping("/api/user")
     public ResponseEntity FindUserWithJwt(HttpServletRequest request){
 
         String authorization = request.getHeader(HttpHeaders.AUTHORIZATION);
@@ -70,5 +67,70 @@ public class UserController {
                 .headers(headers)
                 .body(userDto);
     }
+
+    @GetMapping("/api/users/session")
+    public ResponseEntity getSessionFromCurrentUser() throws IllegalAccessException {
+
+        HttpResponseUserDto userDto = userService.findUserBySession();
+
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body(userDto);
+    }
+
+    @DeleteMapping("/api/users/{userId}")
+    public ResponseEntity DeleteUser(@PathVariable (name = "userId") long userId,HttpServletRequest request) throws IllegalAccessException {
+
+        validateCurrentUser(userId);
+        userService.deleteUser(userId);
+
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .build();
+    }
+
+    @PatchMapping("/api/users/{userId}")
+    public ResponseEntity modifyUserWithJwt(HttpServletRequest request, @Valid @RequestBody HttpRequestModifyUserDto dto, @PathVariable(name = "userId") Long userid) throws IllegalAccessException {
+
+        String authorization = request.getHeader(HttpHeaders.AUTHORIZATION);
+
+        jwtUtil.VerifyToken(authorization,"Identification");
+//        2차 비밀번호 , 시큐리티 레벨이 2인지 검증
+        if(dto.getSecurityState().equals(SecurityState.SECONDPASSWORD) && !dto.getSecondPassword().matches(SECONDPASSWORD_PATTERN)){
+            throw new NotAllowedSecondPasswordException(NOT_ALLOWED_SECONDPASSWORD_LITERAL);
+        }
+        if(!dto.getSecurityState().equals(SecurityState.SECONDPASSWORD) && !(dto.getSecondPassword() == null)){
+            throw new NotAllowedSecondPasswordException(NOT_ALLOWED_SECONDPASSWORD_LITERAL);
+        }
+
+        validateCurrentUser(userid);
+
+        userService.ModifyUser(userid,dto);
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.add(HttpHeaders.AUTHORIZATION,authorization);
+
+        return ResponseEntity.status(HttpStatus.NO_CONTENT)
+                .headers(headers)
+                .body(null);
+    }
+
+    private void validateCurrentUser(long userid) throws IllegalAccessException {
+
+        try {
+            UsmsUserDetails userDetails = (UsmsUserDetails) SecurityContextHolder
+                    .getContext()
+                    .getAuthentication()
+                    .getPrincipal();
+
+            if(userDetails.getId() != userid){
+                throw new IllegalAccessException(NOT_MATCHED_SESSION_USRID_LITERAL);
+            }
+        }
+        catch (ClassCastException | NullPointerException e){
+            throw new IllegalAccessException(NOT_MATCHED_SESSION_USRID_LITERAL);
+        }
+    }
+
 
 }
