+++ src/main/java/com/ssg/usms/business/user/dto/HttpResponseUserDto.java
@@ -7,6 +7,7 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
+@Setter
 @ToString
 public class HttpResponseUserDto {
