+++ src/test/java/com/ssg/usms/business/accident/repository/AccidentRepositoryTest.java
@@ -2,6 +2,7 @@ package com.ssg.usms.business.accident.repository;
 
 import com.ssg.usms.business.accident.constant.AccidentBehavior;
 import com.ssg.usms.business.accident.dto.AccidentStatDto;
+import com.ssg.usms.business.config.EmbeddedRedis;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,7 +19,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @Transactional
 @ActiveProfiles("test")
-@SpringBootTest
+@SpringBootTest(classes = EmbeddedRedis.class)
 public class AccidentRepositoryTest {
 
     @Autowired
