+++ src/test/java/com/ssg/usms/business/cctv/controller/CctvControllerDeletingTest.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ssg.usms.business.cctv.exception.NotExistingCctvException;
 import com.ssg.usms.business.cctv.exception.NotOwnedCctvException;
 import com.ssg.usms.business.cctv.service.CctvService;
+import com.ssg.usms.business.config.EmbeddedRedis;
 import com.ssg.usms.business.error.ErrorResponseDto;
 import com.ssg.usms.business.store.exception.NotExistingStoreException;
 import com.ssg.usms.business.store.exception.NotOwnedStoreException;
@@ -35,7 +36,7 @@ import static org.mockito.BDDMockito.willThrow;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
 
 @ActiveProfiles("test")
-@SpringBootTest
+@SpringBootTest(classes = EmbeddedRedis.class)
 public class CctvControllerDeletingTest {
 
     private MockMvc mockMvc;
