+++ src/test/java/com/ssg/usms/business/config/EmbeddedRedis.java
@@ -2,18 +2,23 @@ package com.ssg.usms.business.config;
 
 import org.springframework.boot.autoconfigure.data.redis.RedisProperties;
 import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
+import org.springframework.session.data.redis.config.annotation.web.http.EnableRedisHttpSession;
 import redis.embedded.RedisServer;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
 
 @TestConfiguration
+@EnableRedisRepositories
 public class EmbeddedRedis {
 
     private RedisServer redisServer;
 
     public EmbeddedRedis(RedisProperties redisProperties) {
-        this.redisServer = new RedisServer(redisProperties.getPort());
+        this.redisServer = RedisServer.builder()
+                .port(redisProperties.getPort())
+                .setting("maxmemory 10M").build();
     }
 
     @PostConstruct
