+++ src/test/java/com/ssg/usms/business/store/repository/StoreRepositoryTest.java
@@ -1,5 +1,6 @@
 package com.ssg.usms.business.store.repository;
 
+import com.ssg.usms.business.config.EmbeddedRedis;
 import com.ssg.usms.business.store.constant.StoreState;
 import com.ssg.usms.business.store.dto.HttpRequestRetrievingStoreDto;
 import com.ssg.usms.business.store.exception.NotExistingStoreException;
@@ -21,7 +22,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @Transactional(readOnly = true)
 @ActiveProfiles("test")
-@SpringBootTest
+@SpringBootTest(classes = EmbeddedRedis.class)
 public class StoreRepositoryTest {
 
     @Autowired
