+++ src/test/java/com/ssg/usms/business/user/UserServiceTest.java
@@ -1,11 +1,13 @@
 package com.ssg.usms.business.user;
 
 
-import com.ssg.usms.business.user.dto.HttpRequestModifyUserDto;
-import com.ssg.usms.business.user.dto.HttpResponseUserDto;
-import com.ssg.usms.business.user.dto.SecurityState;
+import com.ssg.usms.business.security.login.UsmsUserDetails;
+import com.ssg.usms.business.user.dto.*;
+import com.ssg.usms.business.user.exception.AlreadyExistIdException;
+import com.ssg.usms.business.user.exception.AlreadyExistPhoneNumException;
 import com.ssg.usms.business.user.exception.NotExistingUserException;
 import com.ssg.usms.business.user.repository.UserRepository;
+import com.ssg.usms.business.user.repository.UserSessionRepository;
 import com.ssg.usms.business.user.repository.UsmsUser;
 import com.ssg.usms.business.user.service.UserService;
 import com.ssg.usms.business.user.util.JwtUtil;
@@ -18,18 +20,31 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
+import java.util.List;
 import java.util.Optional;
 
+import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
 public class UserServiceTest {
     @Mock
     private UserRepository repository;
+
+    @Mock
+    private UserSessionRepository userSessionRepository;
 
     @Mock
     private BCryptPasswordEncoder bCryptPasswordEncoder;
@@ -38,6 +53,69 @@ public class UserServiceTest {
     private JwtUtil jwtUtil;
     @InjectMocks
     private UserService userService;
+
+    @DisplayName("중복된 아이디를 요청한 경우 AlreadyExistIdException 을 던진다.")
+    @Test
+    public void testSinupWithDupicateId() {
+
+        given(repository.existsByUsername(any())).willReturn(true);
+
+        HttpRequestSignUpDto dto = new HttpRequestSignUpDto();
+        dto.setUsername("httpRequestSign");
+        dto.setPassword("hashedpassword123@");
+        dto.setPhoneNum("010-1234-24124");
+        dto.setEmail("asdf123@naer.com");
+        dto.setNickname("hihello");
+
+
+        assertThrows(AlreadyExistIdException.class, () -> userService.SignUp(dto));
+    }
+
+    @DisplayName("중복된 전화번호를 요청한 경우 AlreadyExistPhoneNumbException 을 던진다..")
+    @Test
+    public void testSinupWithDupicatePhoneNum() {
+
+        given(repository.existsByPhoneNumber(any())).willReturn(true);
+
+        HttpRequestSignUpDto dto = new HttpRequestSignUpDto();
+        dto.setUsername("httpRequestSign");
+        dto.setPassword("hashedpassword123@");
+        dto.setPhoneNum("010-1234-24124");
+        dto.setEmail("asdf123@naer.com");
+        dto.setNickname("hihello");
+
+
+        assertThrows(AlreadyExistPhoneNumException.class, () -> userService.SignUp(dto));
+    }
+
+    @DisplayName("올바른 값을 요청한 경우 200 ResponseEntity를 반환한다.")
+    @Test
+    public void testSinupWithValidParam() {
+
+        UsmsUser user = UsmsUser.builder()
+                .username("httpRequestSign")
+                .password("hashedpassword123@")
+                .personName("hihello")
+                .phoneNumber("010-1234-24124")
+                .email("asdf123@naer.com")
+                .id(1L)
+                .build();
+
+        HttpRequestSignUpDto dto = new HttpRequestSignUpDto();
+        dto.setUsername("httpRequestSign");
+        dto.setPassword("hashedpassword123@");
+        dto.setPhoneNum("010-1234-24124");
+        dto.setEmail("asdf123@naer.com");
+        dto.setNickname("hihello");
+        // then
+        assertThatCode(() -> userService.SignUp(dto))
+                .doesNotThrowAnyException();
+    }
+
+
+
+
+
 
     @DisplayName("인자로 들어온 token값으로 성공적으로 유저를 찾아서 HttpResponseDto를 리턴")
     @Test
@@ -267,5 +345,83 @@ public class UserServiceTest {
         assertThrows(IllegalArgumentException.class , () -> userService.ModifyUser(1L,dto));
     }
 
+
+    @DisplayName("세션이 존재할때 세션을통해서 유저정보를 가져와서 dto를 반환한다.")
+    @Test
+    public void testFindUserdtoWithSession() throws Exception{
+
+        UsmsUser user = UsmsUser.builder()
+                .id(1L)
+                .username("tkfka123")
+                .password("asdfasdf2312*")
+                .email("email@naver.com")
+                .personName("hongu")
+                .phoneNumber("010-1234-4544")
+                .isLock(false)
+                .securityState(SecurityState.BASIC)
+                .role(UserRole.ROLE_STORE_OWNER)
+                .build();
+
+        HttpResponseUserDto dto = HttpResponseUserDto.builder()
+                .id(user.getId())
+                .username(user.getUsername())
+                .personName(user.getPersonName())
+                .phoneNumber(user.getPhoneNumber())
+                .securityState(user.getSecurityState())
+                .email(user.getEmail())
+                .build();
+
+        UsmsUserDetails userDetails = new UsmsUserDetails(user);
+        Authentication authentication = new UsernamePasswordAuthenticationToken(userDetails,null, userDetails.getAuthorities());
+        SecurityContext context = SecurityContextHolder.createEmptyContext();
+        context.setAuthentication(authentication);
+        SecurityContextHolder.setContext(context);
+
+
+        Assertions.assertThat(userService.findUserBySession().toString()).isEqualTo(dto.toString());
+    }
+
+    @DisplayName("세션 없이 유저 정보 조회를 시도할 시 에러 발생")
+    @Test
+    public void testFindUserdtoNoAccessSession(){
+
+        String principal = "aksen";
+        Authentication authentication = new AnonymousAuthenticationToken(principal, principal, List.of(()-> "ROLE_ANONYMOUS"));
+        SecurityContext context = SecurityContextHolder.createEmptyContext();
+        context.setAuthentication(authentication);
+        SecurityContextHolder.setContext(context);
+
+        assertThrows(IllegalAccessException.class, () -> userService.findUserBySession());
+    }
+
+
+    @DisplayName("현재 세션에 해당하는 유저의 세션으로 유저정보 삭제를 요청했을때 세션 삭제 및 유저 삭제")
+    @Test
+    public void testDeleteUser() throws Exception{
+
+        UsmsUser user = UsmsUser.builder()
+                .id(1L)
+                .username("tkfka123")
+                .password("asdfasdf2312*")
+                .email("email@naver.com")
+                .personName("hongu")
+                .phoneNumber("010-1234-4544")
+                .isLock(false)
+                .securityState(SecurityState.BASIC)
+                .role(UserRole.ROLE_STORE_OWNER)
+                .build();
+
+        UsmsUserDetails userDetails = new UsmsUserDetails(user);
+        Authentication authentication = new UsernamePasswordAuthenticationToken(userDetails,null, userDetails.getAuthorities());
+        SecurityContext context = SecurityContextHolder.createEmptyContext();
+        context.setAuthentication(authentication);
+        SecurityContextHolder.setContext(context);
+
+        userService.deleteUser(user.getId());
+
+        verify(userSessionRepository,times(1)).deleteSession(user.getUsername());
+        verify(repository,times(1)).delete(anyLong());
+
+    }
 
 }
